Refactor: move Kind to its own file
